Rename configparser to configloader
