+++ config/configloader/config.go
@@ -12,10 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-// Package configparser implements configuration loading from a config.Parser.
+// Package configloader implements configuration loading from a config.Parser.
 // The implementation relies on registered factories that allow creating
 // default configuration for each type of receiver/exporter/processor.
-package configparser
+package configloader
 
 import (
 	"fmt"
