+++ config/configloader/testdata/duplicate-exporter.yaml
