+++ config/configloader/testdata/duplicate-extension.yaml
