+++ config/configloader/testdata/duplicate-pipeline.yaml
