+++ config/configloader/testdata/duplicate-processor.yaml
