+++ config/configloader/testdata/duplicate-receiver.yaml
