+++ config/configloader/testdata/invalid-exporter-name-after-slash.yaml
