+++ config/configloader/testdata/invalid-exporter-section.yaml
