+++ config/configloader/testdata/invalid-exporter-sub-config.yaml
