+++ config/configloader/testdata/invalid-exporter-type.yaml
