+++ config/configloader/testdata/invalid-extension-name-after-slash.yaml
