+++ config/configloader/testdata/invalid-extension-section.yaml
