+++ config/configloader/testdata/invalid-extension-sub-config.yaml
