+++ config/configloader/testdata/invalid-extension-type.yaml
