+++ config/configloader/testdata/invalid-pipeline-name-after-slash.yaml
