+++ config/configloader/testdata/invalid-pipeline-section.yaml
