+++ config/configloader/testdata/invalid-pipeline-sub-config.yaml
