+++ config/configloader/testdata/invalid-pipeline-type.yaml
