+++ config/configloader/testdata/invalid-processor-name-after-slash.yaml
