+++ config/configloader/testdata/invalid-processor-section.yaml
