+++ config/configloader/testdata/invalid-processor-sub-config.yaml
