+++ config/configloader/testdata/invalid-processor-type.yaml
