+++ config/configloader/testdata/invalid-receiver-name-after-slash.yaml
