+++ config/configloader/testdata/invalid-receiver-section.yaml
