+++ config/configloader/testdata/invalid-receiver-sub-config.yaml
