+++ config/configloader/testdata/invalid-receiver-type.yaml
