+++ config/configloader/testdata/invalid-sequence-value.yaml
