+++ config/configloader/testdata/invalid-service-extensions-section.yaml
