+++ config/configloader/testdata/unknown-exporter-type.yaml
