+++ config/configloader/testdata/unknown-extension-type.yaml
