+++ config/configloader/testdata/unknown-pipeline-type.yaml
