+++ config/configloader/testdata/unknown-processor-type.yaml
