+++ config/configloader/testdata/unknown-receiver-type.yaml
