+++ service/parserprovider/default_test.go
@@ -24,7 +24,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"go.opentelemetry.io/collector/config"
-	"go.opentelemetry.io/collector/config/configparser"
+	"go.opentelemetry.io/collector/config/configloader"
 	"go.opentelemetry.io/collector/processor/attributesprocessor"
 	"go.opentelemetry.io/collector/processor/batchprocessor"
 	"go.opentelemetry.io/collector/receiver/jaegerreceiver"
@@ -49,7 +49,7 @@ func TestDefault(t *testing.T) {
 		require.NoError(t, err)
 		require.NotNil(t, cp)
 		var cfg *config.Config
-		cfg, err = configparser.Load(cp, factories)
+		cfg, err = configloader.Load(cp, factories)
 		require.Error(t, err)
 		require.Nil(t, cfg)
 
@@ -69,7 +69,7 @@ func TestDefault(t *testing.T) {
 		require.NoError(t, err)
 		require.NotNil(t, cp)
 		var cfg *config.Config
-		cfg, err = configparser.Load(cp, factories)
+		cfg, err = configloader.Load(cp, factories)
 		require.NoError(t, err)
 		assert.NotNil(t, cfg)
 		err = cfg.Validate()
@@ -105,7 +105,7 @@ func TestDefault(t *testing.T) {
 		require.NoError(t, err)
 		require.NotNil(t, cp)
 		var cfg *config.Config
-		cfg, err = configparser.Load(cp, factories)
+		cfg, err = configloader.Load(cp, factories)
 		require.NoError(t, err)
 		require.NotNil(t, cfg)
 		err = cfg.Validate()
@@ -145,7 +145,7 @@ func TestDefault_ComponentDoesNotExist(t *testing.T) {
 	cp, err := pl.Get()
 	require.NoError(t, err)
 	require.NotNil(t, cp)
-	cfg, err := configparser.Load(cp, factories)
+	cfg, err := configloader.Load(cp, factories)
 	require.NoError(t, err)
 	require.NotNil(t, cfg)
 }
